Simplify ECCS module layout
